requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1617740417847/work
astroid==2.4.2
autopep8==1.5.5
certifi==2020.12.5
colorama==0.4.4
cycler==0.10.0
dj-database-url==0.5.0
Django @ file:///tmp/build/80754af9/django_1620309509374/work
django-heroku==0.3.1
gunicorn==20.1.0
imutils==0.5.4
isort==5.6.4
joblib==1.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.4.2
mccabe==0.6.1
numpy==1.20.3
opencv-python==4.5.2.52
Pillow==8.2.0
psycopg2 @ file:///C:/ci/psycopg2_1612298766762/work
pycodestyle==2.7.0
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
scikit-learn==0.24.2
scipy==1.6.3
six==1.15.0
sklearn==0.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
threadpoolctl==2.1.0
toml==0.10.2
whitenoise==5.2.0
wincertstore==0.2
wrapt==1.12.1
